add initial type wrappers
This patch adds type conversion wrapper types extracted
from github.com/lib/pq. We test against jackc/pgx/v4/stdlib
as that is the recommended replacement driver for lib/pq.

The tests are all lifted from lib/pq, so some of them test
the driver as well as the type parsers, which isn't really
right, but it is good enough for an initial version. I'll
try to make these tests more focused as problems with them
arise.